requirements.txt
paramiko
pyvmomi
pyyaml
requests
pywinrm
colorama